refactor: Improve session structure in upload image and set VM to n1-standard-4
